fix: removed circular dependencies from defimanagerprovider
