+++ src/features/defi/providers/yearn/components/YearnManager/YearnManager.tsx
@@ -1,7 +1,7 @@
 import { Center, Heading, Stack } from '@chakra-ui/layout'
 import { ModalHeader, useColorModeValue } from '@chakra-ui/react'
 import { DefiActionButtons } from 'features/defi/components/DefiActionButtons'
-import { DefiParams } from 'features/defi/contexts/DefiManagerProvider/DefiManagerProvider'
+import { DefiParams } from 'features/defi/contexts/DefiManagerProvider/DefiTypes'
 import { useYearn } from 'features/defi/contexts/YearnProvider/YearnProvider'
 import { AnimatePresence } from 'framer-motion'
 import { Location } from 'history'
